requirements.txt
flask==2.3.3
python-dotenv==1.0.0
pandas==2.1.0
numpy==1.25.2
openai==1.20.0
nltk==3.8.1
PyPDF2==3.0.1
scikit-learn==1.3.0
tiktoken==0.5.1
langchain==0.1.4
langchain-openai==0.0.4 